--- packages/design-system/src/components/index.ts
@@ -1,1 +1,0 @@
-export * from './button';
