+++ packages/design-system/src/components/ui/button/Button.stories.ts
@@ -2,7 +2,7 @@ import type { Meta, StoryObj } from '@storybook/react';
 import Button from './Button';
 
 const meta: Meta<typeof Button> = {
-  title: 'stories/Button',
+  title: 'components/Button',
   component: Button,
   tags: ['autodocs'],
 };
