+++ packages/design-system/src/components/ui/button/Button.tsx
@@ -1,5 +1,5 @@
-import { ComponentProps, ReactNode } from 'react';
-import { cn } from '../../lib/cn';
+import { ComponentProps } from 'react';
+import { cn } from '../../../lib/utils';
 
 interface ButtonProps extends ComponentProps<'button'> {
   size?: 'small' | 'large';
